Rename slack-adapter
