+++ src/controller.js
@@ -1,6 +1,6 @@
 import { Botkit } from "botkit";
 import { receiveMiddleware, sendMiddleware } from "./middleware";
-import { adapter } from "./slack_adapter";
+import { adapter } from "./slack-adapter";
 
 export const controller = new Botkit({
   webhook_uri: "/api/messages",
