+++ src/slack-adapter.js
